Trying to find who sets server and regionState to null around servercrashprocedure add DEBUG. Ditto for why we do a suspend though we have not done dispatch (on a retry....)
Add to MasterServices ability to wait on Master being up -- makes
it so can Mock Master and start to implement standalone split testing.
Start in on a Split region standalone test in TestAM.

Fix bug where a Split can fail because it comes in in the middle of
a Move (by holding lock for duration of a Move).

+++ b/hbase-server/src/main/java/org/apache/hadoop/hbase/master/assignment/MoveRegionProcedure.java
@@ -61,6 +61,15 @@ public class MoveRegionProcedure
   }